project: add CheckDeleted to handle deleted resource
If a resource is deleted outside of terraform the provide was returning
an error and a terraform rm state was the only way to recover from this.

This commit take a similar approach to what the
terraform-provider-openstack does by adding a CheckDeleted method that
would ignore 404 error on read/delete method and do a setId so that
Terraform understands that the resource have been already deleted.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@cern.ch>