Add fqdn and shortname attributes to Foreman host, handle name
This commit introduces the new attributes "fqdn" and "shortname" to the
Terraform provider for foreman_host objects. Formerly, the "name"
attribute was used to configure and read the hostname of a host. But
this value can be inconsistent, because Foreman might expand a short
hostname to an FQDN, depending on a setting. (this is not a bug)

Because of this unexpected behaviour (results in "inconstistent plan" in
Terraform) the name attribute is now bypassed. Shortname is the new
required input variable.

Summary:
* name: was required, is now optional and computed
* fqdn: new, read-only, always returns the FQDN
* shortname: required and checked for dots (".") to issue warnings

Refs #116