Remove build flag and re-introduce provision method attribute
The build flag was introduced as the replacement for the "method"
attribute. Earlier, the method could be set to "build" or "image", which
is then used by Foreman. The initial creation of a host also needed the
build flag to be set in the API request to signal Foreman that it should
create the machine.

This lead to the problem that Terraform configs containing "build = true"
would reset the build flag in Foreman even for existing, already built
machines, resetting them at the next boot. This is dangerous and should
be avoided.

The commit removes the build flag completely from Terraform and uses the
"provision_method" attribute from the Foreman API (again).

We will have to find an additional attribute to support the use case of
"enforced rebuild" that originally lead to the switch.

The "managed" flag is not changed.

Refs #40
Refs #68
Refs #106