Disable hostname check in resourceForemanHostCreate using settings API
For testing and as an example, the resourceForemanHostCreate used the
newly added settings API to check the hostname of possible
inconsistencies. As @agriffit79 mentioned in the PR #120 the DomainName
value will always be empty / evaluate to false since it is computed.
Therefore, the check is redundant.

It might happen that the DomainName attribute will become available in
other use cases (e.g. directly specified on the host?) - that's why the
check is disabled and not removed completely at the moment.