Fix shortname/fqdn tests; make name argument deprecated and read-only
The tests use an API mock response file to check against a mocked
Terraform schema. Since neither "shortname" nor "fqdn" host attributes
are in the Foreman API, we exclude them in the tests. They are
constructed from the "name" given by Foreman, or passed in in case of
the shortname.

To enforce the switch from "name" to "shortname", name is now a
deprecated argument. It is still accessible via the name attribute,
which is filled when the resource is read from the API.