Re-add ability to set build flag (via set_build_flag argument)
PR #122 removed the "build" argument, removing the possibility to
enforce setting the Foreman-internal build flag to "true". This commit
re-introduces this ability, but with a more explicit argument name:
"set_build_flag". And it defaults to "false", so that not specifying it
does not toggle the flag to "true" on accident.

Resolves #130

See https://github.com/terraform-coop/terraform-provider-foreman/discussions/125