Implement settings API as data source (#120)
* Implement settings API as data source

Implements the API calls for ReadSetting and QuerySetting, allowing a
read-only usage of the Foreman settings API.

As an example, the setting "append_domain_name_for_hosts" is read in the
resourceForemanHostCreate func to handle the issue with short names and
FQDNs as return value.

Refs #116

* Add settings_type to Terraform setting data resource; add example .tf

Adds the settings_type field to the Terraform data resource definition.
See the created example on how to use the setting DR in Terraform.

* Disable hostname check in resourceForemanHostCreate using settings API

For testing and as an example, the resourceForemanHostCreate used the
newly added settings API to check the hostname of possible
inconsistencies. As @agriffit79 mentioned in the PR #120 the DomainName
value will always be empty / evaluate to false since it is computed.
Therefore, the check is redundant.

It might happen that the DomainName attribute will become available in
other use cases (e.g. directly specified on the host?) - that's why the
check is disabled and not removed completely at the moment.

* Settings API: remove TF warnings on type conversion

Refs #120

* Remove commented leftover code for settings API query example

* Add tests for settings API / settings data resource

Adds Terraform schema tests and testdata JSON files for
the Foreman settings API.

The test structure is quite complicated..