Images: fix logic and tests by removing the ".Name" field from ForemanImage
The .Name field/member overshadowed the .Name member of the embedded
ForemanObject struct. Since they are both type string and identified by
the "name" key in JSON, this commit removes the member from the
ForemanImage struct.

This also repairs the DeepEqual test, since .Name was not set in the
embedded struct.