Add compatibility to "name" argument (#132)
* Add compatibility to "name" argument

The breaking change of making host.name read-only renders a lot of TF
configs unusable. This commit adds diff suppression handling to work
with name, shortname and fqdn even with older manifests.

* Remove deprecation of foreman_host.name attribute; extend README

The deprecation was meant to be an improvement but causes breaking
changes to many Terraform manifests. We removed the deprecation and
extended the README instead.