Fix Katello product create/destroy; implement async foreman_tasks
On the way to fix the creation, modification and deletion of Katello
products, it turned out that Katello uses async responses in its API.
Specifically, the destroy (DELETE) endpoints return an async task, which
must be checked for completion. This check is now implemented. If a
product is destroyed, the API client detects the usage of foreman_tasks
and invokes a routine to check the task by its ID, waiting for it to
finish.

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>