job_templates: Fix wrong order of template_inputs from API
The template inputs are a list inside the job_templates objects. After
creating the resources through the Foreman API, reading the job_template
might result in a different order of template_inputs. This causes
Terraform to think there was a change. This commit introduces a sort
function which keeps the template_inputs in order, according to their
unique ID.

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>