katello_repo: Add DiffSuppression to "download_concurrency" parameter.
The "download_concurrency" parameter is not returned from the Katello
API, but still exists in the source code of Katello in v4.9. It's
unclear what happens with the value, if supplied. But as it is not
returned to reflect the Terraform input, TF will always see a change
from 0 -> X. Therefore, DiffSuppression captures theses cases.

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>