katello_sync_plan: Handle "sync_date" diffs better
The "sync_date" field allows passing in a datetime as string. It
specifies the first time a sync plan should run. This field requires a
specific format to allow Go to parse the value.

Use the specified format: YYYY-MM-DD HH:MM:SS +0000, where '+0000' is
the timezone difference. A value of '+0000' means UTC.

Additionally, the time zone suffix was documented as "UTC". There is now
an exception handler for this case, converting "UTC" to "+0000"
internally.

The diff suppression func in this commit allows using shorter fields
like "2023-12-11", which is then parsed to "2023-12-11 00:00:00 +0000".

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>