Fixes for Katello repositories and sync_plans (#143)
* (cherry pick) katello_product: fix product data source query by adding orgId

As seen in
https://apidocs.theforeman.org/katello/latest/apidoc/v2/products.html,
the organization_id is a required parameter for the GET endpoint. It is
now sent with the name query term.


* katello_product: Fix orgID parameter in creation call


* katello_repo: fix "label" field due to missing "computed"

The "label" field was missing the "computed" attribute, marking it as
changed every time if not specified in the .tf file


* katello_repo: replace mirror_on_sync by mirroring_policy; marshalJSON

"mirror_on_sync" is deprecated and removed from Katello (in 4.9 it seems).
It is replaced by "mirroring_policy".

This commit also adds a JSON marshaler to katello_repository for better
handling of fields and use cases (especially the content_type).


* katello_repo: Add DiffSuppression to "download_concurrency" parameter.

The "download_concurrency" parameter is not returned from the Katello
API, but still exists in the source code of Katello in v4.9. It's
unclear what happens with the value, if supplied. But as it is not
returned to reflect the Terraform input, TF will always see a change
from 0 -> X. Therefore, DiffSuppression captures theses cases.


* (cherry pick) Katello product: re-add SSL-related fields


* katello_sync_plan: Handle "sync_date" diffs better

The "sync_date" field allows passing in a datetime as string. It
specifies the first time a sync plan should run. This field requires a
specific format to allow Go to parse the value.

Use the specified format: YYYY-MM-DD HH:MM:SS +0000, where '+0000' is
the timezone difference. A value of '+0000' means UTC.

Additionally, the time zone suffix was documented as "UTC". There is now
an exception handler for this case, converting "UTC" to "+0000"
internally.

The diff suppression func in this commit allows using shorter fields
like "2023-12-11", which is then parsed to "2023-12-11 00:00:00 +0000".


* katello sync_plan: Add ValidateDiagFunc for sync_date value


---------

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>