job_templates: Fix testdata and extend tests for resource
The testdata was missing two fields in the query response: template and
locked. These are not returned in the upstream Foreman API when the
query is used. They are only contained in the detailed GET request for a
specific resource at /api/job_templates/<ID>. This complicates the tests
and therefore the testdata now returns the fields as well, otherwise a
refactoring to split single_query and read would be needed.

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>