Implement utils.Debugf and utils.Fatalf as wrappers
Both functions wrap the log.* functions, adding the filename and line
number of the calling function. Uses a switch to determine the final
log.* function that is called.

jobtemplate and template_input were refactored to use 'Debugf' instead
of 'Debug'. For testing, domain.go was also refactored to use utils.*
instead of log.*.

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>