Fix Katello product resource API call (#142)
* Fix Katello product resource API call JSON payload

foreman_katello_product used an old wrapper that did not include
location_id and organization_id in the payload of calls towards the API.

Resolves #54

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>

* Fix Katello product create/destroy; implement async foreman_tasks

On the way to fix the creation, modification and deletion of Katello
products, it turned out that Katello uses async responses in its API.
Specifically, the destroy (DELETE) endpoints return an async task, which
must be checked for completion. This check is now implemented. If a
product is destroyed, the API client detects the usage of foreman_tasks
and invokes a routine to check the task by its ID, waiting for it to
finish.

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>

* Katello product: re-add SSL-related fields

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>

* katello_product: fix product data source query by adding orgId

As seen in
https://apidocs.theforeman.org/katello/latest/apidoc/v2/products.html,
the organization_id is a required parameter for the GET endpoint. It is
now sent with the name query term.

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>

* Move foreman_task into own api/ file

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>

* Update example for Katello sync_plan

Signed-off-by: Dominik Pataky <pataky@mindful-security.eu>