feat: Implement publish new CV; handle AsyncTask in creation
This commit introduces an additional step, publishing a first version
after creating a new content view. This is needed for creating composite
content views and an experiment.

Sadly, on the way to implement this step the ForemanTask API made
problems again. These are solved by handling the return label of async
tasks and querying the updated resources.

Next steps: add a "latest version" field to the Terraform content view
resource so we can use this ID for "composite_ids". (Which is currently
broken, because the composite_id expects to be a version_id)

Signed-off-by: Dominik Pataky <git@pataky.it>