fix(repositories): Handle download_concurrency missing API return value
Commit introduces multiple fixes to handle the missing
"download_concurrency" parameter in Katello API responses. The func
handleDownloadConcurrencyBetweenTerraformAndKatello basically compares
the API response with the state, so if the API did not return the
value, but we know the value from the Terraform state, we can override
this value in the repository struct.

Also adds validation and a fix for the ContentViewFilter query, which
seems to have been missed in previous test runs.

Resolves https://github.com/terraform-coop/terraform-provider-foreman/issues/161

Signed-off-by: Dominik Pataky <git@pataky.it>