Fixed more tests.
Host tests expect build param to be false.

Host tests expect interface attributes to exist under compute_attributes, and not interfaces.

Operating system tests expect media to be a slice of ints, not a slice of maps.

Subnet tests expect domains to be an empty slice. Added some logging for this.

Parse hostgroup parameters whether they're a slice or a map.