Fix host rebuild (#190) (#191)
* Fix host rebuild (#190). Invoke resourceForemanHostUpdate when build flag drifts from desired state. Also start making UpdateHost API call based on set_build_flag argument. Create test for updating the build flag.

* Resolve set_build_flag test failure and improve attribute comparison

* Fixed more tests.

Host tests expect build param to be false.

Host tests expect interface attributes to exist under compute_attributes, and not interfaces.

Operating system tests expect media to be a slice of ints, not a slice of maps.

Subnet tests expect domains to be an empty slice. Added some logging for this.

Parse hostgroup parameters whether they're a slice or a map.

---------

Co-authored-by: Lennart Weller <lhw@ring0.de>